Create a base class for the databases
